requirements.txt
mysql-connector-python>=8.0.0
pycryptodome>=3.18.0
